complete support for checkboxes on effect controls

+++ Audio_1062_DSP/source/effects_tester/gui/toolkit_gui.h
@@ -14,8 +14,8 @@ typedef struct {
 	EFFECT_PARAMS *myParameter;
 	char *name;
 	uint8_t useCheckBox;
-	uint8_t previousCheck;
-	uint8_t *chkOutput;
+	int previousCheck;
+	int *chkOutput;
 	float slope;
 	float intercept;
 	float *slOutput;
